requirements.txt
requests
mysql-connector-python
matplotlib

Use "pip install -r requirements.txt" to install the requirements